XSpec arg implemented
