--- gui_dev/gui_linelists.ascii
@@ -1,8 +1,0 @@
-Linelist Names
-Eiger_Strong
-LBG
-Gal
-LLS
-LLS Small
-DLA
-atom
