--- gui_dev/prep_guidb.py
@@ -1,33 +1,0 @@
-# -*- coding: utf-8 -*-
-"""prep_guidb.ipynb
-
-Automatically generated by Colaboratory.
-
-Original file is located at
-    https://colab.research.google.com/drive/16DgXX11SuG0Np0Aiu3uzZ0Eyqdvb-gK2
-"""
-
-import astropy.units as u
-import astropy.coordinates import SkyCoord
-import numpy as np
-import pandas as pd
-
-# READ IN INPUT guessed z catalog 
-# Catalog for guessed redshifts
-# assuming the catalog is a 2D dataFrame:
-# | RA (deg) | DEC (deg) | z_guess |
-# | xxx.xxx. | xxx.xxx.  | xxx.xxx |
-
-
-# Prepare and initialize a database dataFrame for GUI
-# Query all fits files in the curret working directory and loop through them to create the following db
-# assuming the database is a 2D dataFrame:
-# | filename | RA | DEC | z | z_err | Confidence | Flag | Linelist | z_guess |
-# only fill out <filename>, <RA>, <DEC>, and <z_guess>
-
-
-
-# loop through the database entries and match coordinates with the catalog.
-# Populate the database with guessed z values.
-
-#save the database
