--- gui_dev/user_manual.txt
@@ -1,66 +1,0 @@
-        MAIN GUI HELP:
-        Left Widget is the absorber manager. Here you can add known absorbers, guess absorbers
-        and plot or hide the transition lines. If an absorber is determined to be incorrect, removing 
-        the abosrber will delete it from the manager and it will not exist in the output csv file once saving
-        
-        The main canvas is where all following keyboard events are tied to. If interacting with another 
-        widget outside of the canvas, must reclick within the canvas to enable the keyboard events
-        
-        The active zabs manager (below canvas) will display what redshift is currently being used
-        for identifying matching transitions. The Catalog button will automatically add this transition 
-        to the absorber manager
-        
-        --------------Keyboard Events----------
-        
-        'r':   resets/clears the axes and replots the spectra
-        'R':   Keeps the spectra active and will remove all lines/text from the canvas
-        't':   Will restrict the ymax of the canvas to the users current mouse height
-        'b':   Restricts ymin of the canvas to current mouse height
-        'S':   Smoothes the spectra
-        'U':   Unsmooth spectra
-        'x':   Sets left x limit (xmin)
-        'X':   Sets right x limit (xmax)
-        ']':   Shifts canvas to the right
-        '[':   Shifts canvas to the left
-        'Y':   User can input their own y limits
-        'E':   Two E keystrokes will compute rest frame equivalent width at a defined region
-        'G':   Three keystrokes to fit a Gaussian profile. [Currently not drawing on the spectrum]
-
-        'H':   Help Window
-        'v':   Opens Separate Vstack GUI for the user to identify detected transitions
-               Vstack commands will be discussed below
-        
-        'j':   Designed to be used by zooming into a small region of the spectra,
-               finding an absortion region, put mouse in the middle of that region.
-               'j' then opens a transition list window for the user to select which
-               ion they believe the abosrber is located. Once a transition is clicked,
-               the active zabs will display what the redshift should be based on the 
-               mouse x location
-               
-       
-        
-        --------Vstack GUI Keyboard Events--------
-        Upon pressing 'v' in main canvas, a separate window will pop up
-        
-        '>':   Shifts page right (if more than one page)
-        '<':   Shifts page left 
-        'w':   Will change the transition flag between detection and Non-detection
-        
-        
-        ------Notes-------
-        Upon hitting 'v', the transition list will be saved. If redoing an analysis to correct
-        or check transitions, it will identify user that the transition list has already been
-        analyzed for the absorber. If continuing it will overwrite the previous results
-        PLEASE NOTE, it will overwrite, not update the previous results.
-        
-        Save: Saving will generate its own file tree for files. Can rename a folder, but
-            do not rename the files otherwise they will not load. Saving can be done without 
-            evaluate linelists for each absorber or with partially evaluated line lists.
-            Saving will access the current working directory, but can change the filepath to any 
-            desired folder.
-            
-            The absorber manager will be saved as a .csv and the linelists saved as a .txt
-            
-        Load: To load, only give the parent folder, which has contents of the .csv and .txt files
-              Loading will repopulate the previously confirmed absorber redshifts, colors, and linelist 
-              used and will plot the detected absorbers as identified during the VStack GUI.
